Re-enabling error checking on reads and writes, more cleanup, still a fair bit to go.

+++ xid_device_driver/interface_xid_con.h
@@ -9,7 +9,7 @@ namespace cedrus
         virtual ~interface_xid_con()
         {}
 
-        virtual bool read( unsigned char *in_buffer, int bytes_to_read, int *bytes_read) = 0;
+        virtual bool read( unsigned char *in_buffer, int bytes_to_read, unsigned long *bytes_read) = 0;
 
     };
 } // namespace cedrus
